Use an XRAnchor as the base for the model's position
This requires the `anchors` XR feature.

This way if the underlying XR space updates it's origin over time, then the model should no longer jump around.
I.e. the pose of the `XRAnchor` will get updated when the `XRSpace` origin changes, in such a way that the anchors's real world location doesn't change.

In case the tracked anchor is lost, the model will be hidden and the recticle is shown again.
Moved the model delete functionality into a `deleteActionCallback` method so it can also be used when the anchor is lost.