Don't try to delete() anchor on sessions ended.
When calling `delete()` on the `XRAnchor` in the `onXRSessionEnded()` callback it seems to cause a page crash in Chrome.
Without the `delete()` this crash doesn't occur.
So since the anchors are probably cleaned up internally already when the `XRSession` ends, we skip calling it here explicitly.
And thus only clear the `anchor` ref.